+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Upcoming
+## Added
+- New header methods
+
 ## [1.1.7] - 2017-07-12
 ## Fixed
 - Fix default property in getProperty()